+++ src/et_models/convert.jl
@@ -11,6 +11,63 @@ import ACEpotentials.Models: LearnableRnlrzzBasis, PolyEnvelope2sX,
 using LinearAlgebra: norm, dot 
 
 
+function convert2et(model)
+   # TODO: add checks that the model we are importing is of the format 
+   #       that we can actually import and then raise errors if not.
+   #       but since we might just drop this import functionality entirely it
+   #       is not so clear we should waste our time on that. 
+
+   # extract species information from the ACE model 
+   rbasis = model.rbasis
+   et_i2z = ChemicalSpecies.(rbasis._i2z)
+
+   # ---------------------------- REMBED
+   # convert the radial basis 
+   et_rbasis = _convert_Rnl_learnable(rbasis) 
+   et_rspec = rbasis.spec
+   # convert the radial basis into an edge embedding layer which has some 
+   # additional logic for handling the ETGraph input correctly 
+   rembed = ET.EdgeEmbed( et_rbasis; name = "Rnl" )
+
+   # ---------------------------- YEMBED 
+   # convert the angular basis
+   ybasis = model.ybasis
+   et_ybasis = ET.EmbedDP( ET.NTtransform(x -> x.𝐫), 
+                           ybasis )
+   et_yspec = P4ML.natural_indices(et_ybasis.basis)
+   yembed = ET.EdgeEmbed( et_ybasis; name = "Ylm" )
+
+   # ---------------------------- MANY-BODY BASIS
+   # Convert AA_spec from (n,l,m) format to (n,l) format for mb_spec
+   AA_spec = model.tensor.meta["𝔸spec"] 
+   et_mb_spec = unique([[(n=b.n, l=b.l) for b in bb] for bb in AA_spec])
+
+   et_mb_basis = ET.sparse_equivariant_tensor(
+         L = 0,  # Invariant (scalar) output only
+         mb_spec = et_mb_spec,
+         Rnl_spec = et_rspec,
+         Ylm_spec = et_yspec,
+         basis = real
+      )
+
+   # ---------------------------- READOUT LAYER
+   # readout layer : need to select which linear operator to apply 
+   # based on the center atom species
+   selector = let zlist = et_i2z
+      x -> ET.cat2idx(zlist, x.z)
+   end
+   readout = ET.SelectLinL(
+                  et_mb_basis.lens[1],  # input dim (mb basis length)
+                  1,                    # output dim (only one site energy per atom)
+                  length(et_i2z),       # number of categories = num species 
+                  selector)            
+
+   # generate the model and return it 
+   et_model = ETACE(rembed, yembed, et_mb_basis, readout)
+   return et_model
+end
+
+
 # In ET we currently store an edge xij as a NamedTuple, e.g, 
 #    xij = (𝐫ij = ..., zi = ..., zj = ...)
 # The NTtransform is a wrapper for mapping xij -> y
