+++ src/et_models/et_ace.jl
@@ -25,72 +25,58 @@ function _apply_etace(l::ETACE, X::ET.ETGraph, ps, st)
    Ylm, _ = l.yembed(X, ps.yembed, st.yembed)
 
    # many-body basis 
-   𝔹, _ = l.basis((Rnl, Ylm), ps.basis, st.basis)
+   (𝔹,), _ = l.basis((Rnl, Ylm), ps.basis, st.basis)
 
    # readout layer 
-   φ, _ = l.readout((𝔹[1], X.node_data), ps.readout, st.readout)
+   φ, _ = l.readout((𝔹, X.node_data), ps.readout, st.readout)
 
    # TODO: return site energies or total energy? 
    #       for THIS layer probably site energies, then write all 
    #       the summation and differentiation in the calculator layer. 
-   #       so this is only temporary for testing. 
 
-   return sum(φ)
+   return φ
+end
+
+# ----------------------------------------------------------- 
+
+import Zygote 
+
+#
+# At first glance this looks like we are computing ∂E / ∂ri but this is not 
+# actually true. Because E = ∑ Ei and by interpreting G as a list of edges 
+# we are differentiating E w.r.t. 𝐫ij which is the same is Ei w.r.t. 𝐫ij.
+#
+
+function site_grads(l::ETACE, X::ET.ETGraph, ps, st)
+   ∂X = Zygote.gradient( X -> sum(_apply_etace(l, X, ps, st)), X)[1]
+   return ∂X
 end
 
 
-function convert2et(model)
-   # TODO: add checks that the model we are importing is of the format 
-   #       that we can actually import and then raise errors if not.
-   #       but since we might just drop this import functionality entirely it
-   #       is not so clear we should waste our time on that. 
+# ----------------------------------------------------------- 
+#    basis and jacobian evaluation 
 
-   # extract species information from the ACE model 
-   rbasis = model.rbasis
-   et_i2z = ChemicalSpecies.(rbasis._i2z)
+#=
+function eval_basis(l::ETACE, X::ET.ETGraph, ps, st)      
+   # embed edges 
+   Rnl, _ = l.rembed(X, ps.rembed, st.rembed)
+   Ylm, _ = l.yembed(X, ps.yembed, st.yembed)
 
-   # ---------------------------- REMBED
-   # convert the radial basis 
-   et_rbasis = _convert_Rnl_learnable(rbasis) 
-   et_rspec = rbasis.spec
-   # convert the radial basis into an edge embedding layer which has some 
-   # additional logic for handling the ETGraph input correctly 
-   rembed = ET.EdgeEmbed( et_rbasis; name = "Rnl" )
+   # many-body basis 
+   𝔹, _ = l.basis((Rnl, Ylm), ps.basis, st.basis)
 
-   # ---------------------------- YEMBED 
-   # convert the angular basis
-   ybasis = model.ybasis
-   et_ybasis = ET.EmbedDP( ET.NTtransform(x -> x.𝐫), 
-                           ybasis )
-   et_yspec = P4ML.natural_indices(et_ybasis.basis)
-   yembed = ET.EdgeEmbed( et_ybasis; name = "Ylm" )
-
-   # ---------------------------- MANY-BODY BASIS
-   # Convert AA_spec from (n,l,m) format to (n,l) format for mb_spec
-   AA_spec = model.tensor.meta["𝔸spec"] 
-   et_mb_spec = unique([[(n=b.n, l=b.l) for b in bb] for bb in AA_spec])
-
-   et_mb_basis = ET.sparse_equivariant_tensor(
-         L = 0,  # Invariant (scalar) output only
-         mb_spec = et_mb_spec,
-         Rnl_spec = et_rspec,
-         Ylm_spec = et_yspec,
-         basis = real
-      )
-
-   # ---------------------------- READOUT LAYER
-   # readout layer : need to select which linear operator to apply 
-   # based on the center atom species
-   selector = let zlist = et_i2z
-      x -> ET.cat2idx(zlist, x.z)
-   end
-   readout = ET.SelectLinL(
-                  et_mb_basis.lens[1],  # input dim (mb basis length)
-                  1,                    # output dim (only one site energy per atom)
-                  length(et_i2z),       # number of categories = num species 
-                  selector)            
-
-   # generate the model and return it 
-   et_model = ETACE(rembed, yembed, et_mb_basis, readout)
-   return et_model
+   return 𝔹[1] 
 end
+
+
+function jacobian_basis(l::ETACE, X::ET.ETGraph, ps, st)      
+   # embed edges 
+   Rnl, _ = l.rembed(X, ps.rembed, st.rembed)
+   Ylm, _ = l.yembed(X, ps.yembed, st.yembed)
+
+   # many-body basis jacobian 
+   (𝔹,), ∂𝔹 = l.basis.jacobian((Rnl, Ylm), ps.basis, st.basis)
+
+   return 𝔹[1], ∂𝔹[1]
+end
+=#
