+++ src/et_models/et_models.jl
@@ -3,7 +3,7 @@ module ETModels
 
 include("et_ace.jl")
 
-include("Rnl_learnable_new.jl")
+include("convert.jl")
 
 
 end
